requirements.txt
requests>=2.31.0
PyGithub>=1.59.0
python-dotenv>=1.0.0
click>=8.1.0
gitpython>=3.1.40